backend/requirements.txt
# recommendation system dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# auth dependencies
passlib>=1.7.4
argon2-cffi>=21.2.0
python-jose>=3.3.0
python-multipart
